Inserting text to display level (for example)
Progress, but still encountering major issues:

The collisions with the door seem to be off because the levels that getting printed only appear with the left of the player's hitbox, where it originally was before having it moved.
The same is for the collisions with the blocks. The collisions with the blocks are not complete, so it is hard to tell whether it actually is broken or not.