Same as before
Notice:
The player hitbox has been modified by the stupidest way... It way an image problem --> resolved by cropping the image to the correct dimensions and removed the function player.hitbox() because it caused to much problems. Now if I need to change the  hitbox, I can manually modify it where I define the player itself.