Same as before
Notice:
there was an issue with the hitbox and I couldn't understand the problem but it was stupidly resolved. I just cropped the image because the image was too wide. Because I cropped the image, the image was smaller and resulting, I was able to delete the player.hitbox() function because it has no more use. I can manually modify the hitbox where I define the player's length and width. Collisions are looking good (for now).