Changed the hitbox of the player on the x and y axis
Why?
Because Python might be a very simple programming language but runs slower than a lot of other programming languages. So when I say that the player image should not be greater than 400, it takes a lot of frames to detect it. Thant's why I am changing the position of the hitbox to make it more easier to detect the collisions.

Now what remains is to change the part where the player is kind of going into the ground because it takes too much time for Python to recognize it.