Made modifications back the original hitbox code
I found out that the modifications I had made were useless:

So I've run my game and I was searching for a possible cause of the player sinking into the ground problem.

This is what I've obtained after running the game from the print methods:
329  = PLAYERS X 395  = PLAYERS Y
210  = BLOCKS X 300  = BLOCKS Y
The return value is []
<Player Sprite(in 1 groups)> <Group(0 sprites)>
The player is at  329 400
Player rectangle is at  329 395
Block is at  210 300
0
3
0
329  = PLAYERS X 416  = PLAYERS Y
210  = BLOCKS X 300  = BLOCKS Y
Player going down!!!
1088 576
The return value is []
<Player Sprite(in 1 groups)> <Group(0 sprites)>
The player is at  329 400
Player rectangle is at  329 400
Block is at  210 300
0
3
0

(This is when the player finishes its jump and is landing on the ground...)

Now what is interesting is that the maximum is 400 and we can observe in the first "loop" that the rectangle of the player (I think it is the hitbox) is at y = 395 pixels and and the player image is at y = 400. Technically the player should be stopped at 400 and the rectangle will join it at an interval of 5 pixels... But what we observe in the second "loop" is that because the rectangle hasn't yet reached the ground at y = 400, then the player is able to slip through the ground until the rectangle arrives on the ground. But what is weird and I don't understand is that the player image sinks further until y = 425, in an interval of many loops in a an increment of one before coming to y = 400. This shows that the hitbox changes were useless because the hotbox was suppose to be greater but it still gave the same result.