Trying to fix the jumping issue (issue found)
Found:
    # Forces the player to have gravity applied to it
    #if player.y < min_y:
    #    player.y += 5 * y_gravity

(lines 544 - 546)

These lines purpose are to add an additional gravity to the player when he jumps off or runs off the blocks and not make him float there in the sky.
But the issue with this code is that it forces the player to constantly have a gravity applied to it and thus have the player sink every time he lands. 

Need:
Find an alternative to these two problems.