Further problem with this code
Bullet:
The code stats that the bullet always departs from a certain x and y level and restarts its motion after colliding with either the block and the player. 
The issue is that the bullet continuously is at the x and y level from where it starts. So if the player somehow is able to collide in any other level the place where the bullet is "created", then he automatically is collided and his life drops.