Added if statement to collision to stop player from moving to side after colliding up and down
Added an if statement that would prevent the player from moving to corners when colliding up and down.
Also removed some useless code.

Small issue:
When the player collides to the right and left, if side keys are pressed, he gradually gets down, like in an elevator. If the keys are not pressed, then the player sticks to side of the block.