Saving before breaking the code
Goals for the next parts:
Making the game over screen as a level that when the lives are 0, the player will be teleported to that level that will have the screen.

Why:
Currently the method I am using is that I freeze the screen but then the program is very bulky I cannot get the program to be smooth. 

Advantages of new changes:
I can do a play again method and reroll the game. And be smooth